Refactored the mod a little
