+++ src/main/java/mcvmcomputers/client/gui/GuiFocus.java
@@ -6,8 +6,8 @@ import java.util.ArrayList;
 import org.lwjgl.BufferUtils;
 import org.lwjgl.glfw.GLFW;
 
-import mcvmcomputers.ClientMod;
 import mcvmcomputers.MainMod;
+import mcvmcomputers.client.ClientMod;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.text.TranslatableText;
