+++ src/main/java/mcvmcomputers/client/utils/VMSettings.java
@@ -1,0 +1,5 @@
+package mcvmcomputers.client.utils;
+
+public class VMSettings {
+
+}
