+++ src/main/java/mcvmcomputers/entities/EntityDeliveryChest.java
@@ -2,8 +2,8 @@ package mcvmcomputers.entities;
 
 import java.util.UUID;
 
-import mcvmcomputers.ClientMod;
 import mcvmcomputers.MainMod;
+import mcvmcomputers.client.ClientMod;
 import mcvmcomputers.item.ItemPackage;
 import mcvmcomputers.utils.TabletOrder;
 import mcvmcomputers.utils.TabletOrder.OrderStatus;
