+++ src/main/java/mcvmcomputers/entities/EntityItemPreview.java
@@ -1,6 +1,6 @@
 package mcvmcomputers.entities;
 
-import mcvmcomputers.ClientMod;
+import mcvmcomputers.client.ClientMod;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.data.DataTracker;
