+++ src/main/java/mcvmcomputers/item/ItemPCCaseSidepanel.java
@@ -2,7 +2,7 @@ package mcvmcomputers.item;
 
 import java.util.List;
 
-import mcvmcomputers.ClientMod;
+import mcvmcomputers.client.ClientMod;
 import mcvmcomputers.entities.EntityPC;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.entity.player.PlayerEntity;
