+++ src/main/java/mcvmcomputers/item/ItemWallTV.java
@@ -1,6 +1,6 @@
 package mcvmcomputers.item;
 
-import mcvmcomputers.ClientMod;
+import mcvmcomputers.client.ClientMod;
 import mcvmcomputers.entities.EntityWallTV;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
