+++ src/main/java/mcvmcomputers/mixins/KeyboardMixin.java
@@ -1,6 +1,6 @@
 package mcvmcomputers.mixins;
 
-import static mcvmcomputers.ClientMod.*;
+import static mcvmcomputers.client.ClientMod.*;
 
 import org.lwjgl.glfw.GLFW;
 import org.spongepowered.asm.mixin.Mixin;
@@ -8,11 +8,11 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import mcvmcomputers.ClientMod;
 import mcvmcomputers.MainMod;
+import mcvmcomputers.client.ClientMod;
 import mcvmcomputers.client.gui.GuiFocus;
 import mcvmcomputers.client.gui.setup.pages.SetupPageUnfocusBinding;
-import mcvmcomputers.utils.KeyConverter;
+import mcvmcomputers.client.utils.KeyConverter;
 import net.minecraft.client.Keyboard;
 import net.minecraft.client.MinecraftClient;
